move CLIParsable into BattleWrapper.Config
